requirements.txt
flask>=3.0.0
sentence-transformers>=2.2.0
openai>=1.3.0
numpy>=1.21.0
scikit-learn>=1.2.0
pandas>=1.5.0
plotly>=5.0.0
python-dotenv>=1.0.0
requests>=2.28.0
nomic>=1.1.0 